test: fix test failures due to prompt not fitting in the window
An unknown upstream change has caused the snacks prompt to contain the
full path to the current working directory. This makes the prompt too
wide to include the old `🔎` search icon, which the tests were relying
on to detect when the search prompt is visible.

Look for the name of the project instead to determine when the search is
ready to be used.